OK the iOS build is ugh.
The DatePicker lib I was using had really weird issues on iOS,
returning data not matching the data in the date picker sometimes.

Tried replacing it with barebones DatePickerIOS elements shown on iOS
and using the library on Android, that worked but those DatePickerIOS
elements are hidable by default so it was messy.

Brought in a different react native library, which works off buttons
showing a modal. That modal is returning weird data too (e.g. the date
is 03/14/1992 but when I hit Confirm it slides the datepicker as the
thing animates off the page, and returns 03/15/1996).

Going to just replace it with the barebones DatePickerIOS tomorrow.